chore: removed `this` usage within ddlProvider context
